chore: reorganize backend folder structure — Phase 4 EE mirror (authorization functions)
Move EE authorization functions from ee/app/authorization/server/ to
ee/server/lib/authorization/ (staying inside ee/), updating import
paths and the jest testMatch glob for validateUserRoles.spec.ts.

callback.ts (hook, Phase 6b) and index.ts stay behind.
