+++ apps/meteor/ee/server/lib/authorization/resetEnterprisePermissions.ts
@@ -1,6 +1,6 @@
 import { Permissions } from '@rocket.chat/models';
 
-import { guestPermissions } from '../lib/guestPermissions';
+import { guestPermissions } from '../../../app/authorization/lib/guestPermissions';
 
 export const resetEnterprisePermissions = async function (): Promise<void> {
 	await Permissions.updateMany(
